faq_viana/requirements.txt
Flask
Flask-Cors
psycopg2-binary
langdetect
faiss-cpu
python-docx